Extracted some private classes into separate files
