+++ src/WebsiteProvider/Api/PinningHandlers.cs
@@ -10,7 +10,7 @@ namespace WebsiteProvider.Api
     /// <summary>
     /// Do mapping/unmapping of the Pinning Rules
     /// </summary>
-    public class PinningHandlers
+    public partial class PinningHandlers
     {
         private static PinningHandlers instance;
         private bool isRegistered;
@@ -261,62 +261,6 @@ namespace WebsiteProvider.Api
             if (!isExpectedRegistered && isRegistered)
             {
                 throw new InvalidOperationException("Mappings is already registered.");
-            }
-        }
-
-        /// <summary>
-        /// Summary info for Pinning Rule mapping
-        /// </summary>
-        private class PinningRuleMappingInfo
-        {
-            public List<ulong> MapIds { get; set; }
-            public ulong SectionId { get; set; }
-            public ulong TemplateId { get; set; }
-
-            public bool UseCustomCatchingRule { get; set; }
-
-            public string CallerUri { get; set; }
-            public string ForeignUri { get; set; }
-            public string Token { get; set; }
-            public string TokenBase { get; set; }
-        }
-
-        private class PinningRulesMappingCollection
-        {
-            private readonly object locker = new object();
-            private readonly List<PinningRuleMappingInfo> mappingInfos = new List<PinningRuleMappingInfo>();
-
-            public void Add(PinningRuleMappingInfo info)
-            {
-                lock (locker)
-                {
-                    this.mappingInfos.Add(info);
-                }
-            }
-
-            public List<PinningRuleMappingInfo> Take(Func<PinningRuleMappingInfo, bool> predicate)
-            {
-                lock (locker)
-                {
-                    var items = this.mappingInfos.Where(predicate).ToList();
-                    foreach (var info in items)
-                    {
-                        if (info.MapIds.Count <= 1)
-                        {
-                            this.mappingInfos.Remove(info);
-                        }
-                    }
-                    return items;
-                }
-            }
-
-            public void Process(Func<PinningRuleMappingInfo, bool> predicate, Action<List<PinningRuleMappingInfo>> action)
-            {
-                lock (locker)
-                {
-                    var items = this.mappingInfos.Where(predicate).ToList();
-                    action(items);
-                }
             }
         }
     }
